Simple fake pinpad that returns a hardcoded value.
The hardcoded value returned is either the initial user or admin pin,
depending on which is being asked for.

If you change a pin, you may need to take advantage of gnupg's magic in
the 'login' DO (append '\n\x14P=n,m\n' where n is the length of the user
pin, m is the length of the admin pin, and set to 0 means don't use the
pinpad for that pin).